+++ JudoPayDotNet/Clients/IPreAuths.cs
@@ -29,7 +29,14 @@ namespace JudoPayDotNet.Clients
         /// </summary>
         /// <param name="pkPreAuth">The apple pay pre authorization.</param>
         /// <returns>The receipt for the created apple pre authorization</returns>
-        Task<IResult<ITransactionResult>> Create(PKPaymentModel pkPreAuth);
+        Task<IResult<ITransactionResult>> Create(ApplePayPaymentModel pkPreAuth);
+
+        /// <summary>
+        /// Creates the specified Google Pay pre authorization.
+        /// </summary>
+        /// <param name="googlePayPayment">The Google Pay pre authorization.</param>
+        /// <returns>The transaction result for the created Google Pay pre authorization</returns>
+        Task<IResult<ITransactionResult>> Create(GooglePayPaymentModel googlePayPayment);
     }
     // ReSharper restore UnusedMember.Global
 }